Got VAE working properly with correct image feed

+++ lib/datasets/mnist.py
@@ -19,13 +19,15 @@ from model.config import cfg
 import numpy as np
 import time
 
+import matplotlib.pyplot as plt
+
 
 class Mnist(object):
     """Fast R-CNN data layer used for training."""
 
     def __init__(self, random=False):
         """Set the roidb to be used by this layer during training."""
-        self._images = np.expand_dims(np.array([i.reshape(28, 28).astype(np.int32) for i in mnist.train.images]), 3)
+        self._images = np.expand_dims(np.array([i.reshape(28, 28) for i in mnist.train.images]), 3)
         self._labels = mnist.train.labels.astype(np.int32)
 
         self._num_train_images = 55000
